+++ routes/api/auth/POST-login.js
@@ -1,0 +1,8 @@
+const login = require('../../../middleware/login');
+const jwtSign = require('../../../middleware/jwtSign');
+
+module.exports = (app) => {
+  app.post('/api/auth/login', login, jwtSign, (req, res) => {
+    res.status(200).json({ message: 'Login Successful!' });
+  });
+};
