+++ routes/device/GET-info.js
