+++ routes/device/POST-register.js
